[MOVE] Moved code to ol_purchase module
