+++ ls_pdf_reports/models/__init__.py
@@ -2,6 +2,5 @@ from . import tools
 from . import mrp
 # TODO: "helpdesk.rma" model not found
 # from . import rma
-from . import purchase
 from . import report_extra_content
 from . import report
